fix(bench): uint8 underflow in --index query construction
np.unpackbits returns uint8, so `[:, :bit] * 2 - 1` computed 0*2-1 = 255
(underflow) instead of -1. pack_codes then thresholds (>0), turning every
real-index query into an all-ones code → garbage results (Recall@10 ~0.002
on real DGX codes). Cast to int8 before the shift.

Caught by running the benchmark against a real trained checkpoint on DGX
Spark (5K COCO test codes): after the fix, 64-bit Recall@10 vs exact float
cosine = 0.366 (was 0.002). The codes themselves are healthy (5000/5000
unique, bit-entropy 0.987). Synthetic mode used int8 codes and dodged this.

Adds a regression test (unpacked query must round-trip to find itself at
distance 0).

+++ scripts/bench_search.py
@@ -106,7 +106,8 @@ def main() -> None:
         rng = np.random.default_rng(args.seed)
         q_idx = rng.choice(len(emb), size=min(args.queries, len(emb)), replace=False)
         hidx = HammingIndex(packed, n_bits=bit)
-        q_codes = np.unpackbits(packed[q_idx], axis=1)[:, :bit] * 2 - 1
+        # uint8 unpack → cast before *2-1, else 0*2-1 underflows to 255 (not -1)
+        q_codes = np.unpackbits(packed[q_idx], axis=1)[:, :bit].astype(np.int8) * 2 - 1
         n = len(emb)
     else:
         args.synthetic = True
